Moved get_cdist in the SiteCollection
