+++ openquake/hazardlib/calc/filters.py
@@ -18,12 +18,11 @@
 
 import ast
 import sys
-import time
 import logging
 import operator
 from contextlib import contextmanager
 import numpy
-from scipy.spatial import cKDTree, distance
+from scipy.spatial import cKDTree
 
 from openquake.baselib.python3compat import raise_
 from openquake.hazardlib import site
@@ -385,19 +384,6 @@ class SourceFilter(object):
         if len(split) > 1:  # also set .nsites on the original source
             src.nsites = len(self.close_sids(src)) or .01
         return split
-
-    # used for debugging purposes and in a test
-    def get_cdist(self, rec_or_loc):
-        """
-        :param rec_or_loc: a record with field 'hypo' or a Point instance
-        :returns: array of N euclidean distances from rec['hypo']
-        """
-        try:
-            lon, lat, dep = rec_or_loc['hypo']
-        except TypeError:
-            lon, lat, dep = rec_or_loc.x, rec_or_loc.y, rec_or_loc.z
-        xyz = spherical_to_cartesian(lon, lat, dep).reshape(1, 3)
-        return distance.cdist(self.sitecol.xyz, xyz)[:, 0]
 
     def filter(self, sources):
         """
